[mod] #214 탈퇴 믹스패널 이벤트 로직 이동
